requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-cors-headers>=2.4.0
graphene-django>=2.2.0
django-filter>=2.0.0
django-graphql-jwt>=0.1.5